Introduce a workaround (a hack really) for ScAbstractType conformance
#SCL-11320
This is a temporary solution, having both ScAbstractType and
UndefinedType makes it hard to follow scalac in how conformance is
meant to work. Also, when performing conformance checks involving both
(e.g. G_[C_] <:< NotInferredA) we frequently end up with constraints,
which are not well formed (e.g. NotInferredA <: Any[_]).